Compile and Test assessment #7 #10 (#11)
* Feat build and test 

The CI server now builds the cloned repository. So it compiles it and runs the unit tests.
If the build is unsuccesful the boolean success is set to false.
Github webhook timeout after 10 sec. Implements multi-threading
to send the response before building the project.
GitUtil class contains static functions to maniupulate the git repo.
New tests for gitClone() and deleteRepo().
Tests that the build is successful for the branch assessment
and fails for the branch assess_error.
Closes #7 
Closes #10