Completes build log functionality on localhost #3
BuildLogger now creates a folder build_history in the working directory
of the server.

Buildlogger creates an html file index, where each build entry is added
to the index with a link to a separate html where information on the
build is presented.

Added test code to CI.java to test the creation of files.

Files remain on the servers working directory if the server stops
running.